Renames the module to chamber access.
Now the module chamber access will have several classes in its service.
For example, you will be able to import ExperimentalAccess from
chamber.access.chamber.service.
